add do file resolve function

+++ md5.h
@@ -5,7 +5,7 @@
 
 typedef struct
 {
-    uint64_t count[2];
+    uint32_t count[2];
     uint32_t state[4];
     uint8_t buffer[64];
 } MD5_CTX;
@@ -45,7 +45,7 @@ void MD5Init(MD5_CTX *context);
 void MD5Update(MD5_CTX *context, unsigned char *input, unsigned int inputlen);
 void MD5Final(MD5_CTX *context, unsigned char digest[16]);
 void MD5Transform(unsigned int state[4], unsigned char block[64]);
-void MD5Encode(unsigned char *output, unsigned int *input, unsigned int len);
+void MD5Encode(unsigned char *output, uint32_t *input, unsigned int len);
 void MD5Decode(unsigned int *output, unsigned char *input, unsigned int len);
 
 #endif
